GhgSatApp/app/requirements.txt
alabaster==0.7.12
Babel==2.7.0
click==6.7
colorama==0.4.1
coverage==4.5.4
docutils==0.15.2
Flask==0.12.3
gunicorn==19.7.1
imagesize==1.1.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nose2==0.9.1
packaging==19.2
Pillow==6.2.1
Pygments==2.4.2
pyparsing==2.4.5
pytz==2017.2
requests==2.13.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
Werkzeug==0.12.1
